Solver for first option with all letters
It will find a word with scrambled letters, only if the correct one is the first one with all the letters.
so let's say the word i want is ZEBRA, if ARBEZ comes in front of ZEBRA in the word list the function will return ARBEZ.
So in conclusion it is probably not effective in an actual wordle game.